+++ src/constants.js
@@ -10,6 +10,7 @@ const DECLARED_LOST = 'Declared lost';
 
 export const BROWSE_INVENTORY_ROUTE = '/inventory/browse';
 export const INVENTORY_ROUTE = '/inventory';
+export const DATA_IMPORT_JOB_PROFILES_ROUTE = 'data-import-profiles/jobProfiles';
 
 export const searchModeSegments = {
   search: 'search',
@@ -208,6 +209,12 @@ export const SORT_DIRECTION = {
   ASCENDING: 'ascending',
   DESCENDING: 'descending',
 };
+
+export const SORTABLE_SEARCH_RESULT_LIST_COLUMNS = {
+  TITLE: 'title',
+  CONTRIBUTORS: 'contributors',
+};
+
 
 export const DEFAULT_FILTERS_NUMBER = 6;
 
@@ -427,6 +434,6 @@ export const SINGLE_ITEM_QUERY_TEMPLATES = {
   'items.barcode': 'barcode==%{query}',
   isbn: 'isbn==%{query}',
   issn: 'issn==%{query}',
-  hrid: 'hrid==%{query}',
+  itemHrid: 'hrid==%{query}',
   iid: 'id==%{query}',
 };